MF-991 - Move ws helpers to transport file
